fix: Move updatefollowlist to followlistItems Component
