+++ fhf-datafangst-client/src/containers/BenchmarkView/BenchmarkView.tsx
@@ -10,7 +10,6 @@ import { FC, useEffect } from "react";
 import { useAuth } from "oidc-react";
 import {
   MenuViewState,
-  clearBenchmarkData,
   getBenchmarkData,
   getTrips,
   selectBenchmarkNumHistoric,
@@ -25,7 +24,7 @@ import {
   useAppDispatch,
   useAppSelector,
 } from "store";
-import { Ordering, TripSorting, Vessel } from "generated/openapi";
+import { Ordering, TripSorting } from "generated/openapi";
 import { GridContainer, HeaderButtonCell, HeaderTrack } from "containers";
 import { ArrowBackIos } from "@mui/icons-material";
 import { useNavigate } from "react-router-dom";
@@ -78,21 +77,6 @@ export const BenchmarkView: FC = () => {
   const navigate = useNavigate();
 
   const followVessels = user?.following.map((id) => fiskeridirVessels[id]);
-
-  const updateFollowVessels = (vessel: Vessel, isFollowing?: number) => {
-    if (!isFollowing) {
-      dispatch(
-        getBenchmarkData({
-          vessels: [vessel],
-          sorting: [TripSorting.StopDate, Ordering.Desc],
-          limit: benchmarkHistoric,
-          offset: 0,
-        }),
-      );
-    } else {
-      dispatch(clearBenchmarkData(vessel));
-    }
-  };
 
   useEffect(() => {
     dispatch(setViewState(MenuViewState.Benchmark));
@@ -177,7 +161,7 @@ export const BenchmarkView: FC = () => {
             <Typography color={"white"} variant="h5" sx={{ padding: "10px" }}>
               FØLGELISTE
             </Typography>
-            <FollowList onChange={updateFollowVessels} />
+            <FollowList />
           </Drawer>
         </FollowerArea>
         <GridMainArea>
